chore: Remove DataBrokerSubscriber
Solution is to sophisticated for the kuksa-java-sdk.
However the managemend of the subscription could be put
inside velocitas vehicle-app-java-sdk at some later point in time.

BREAKING CHANGE: DataBrokerConnection#unsubscribe was removed
the unsubscribe method now returns a CancelableContext which
must be stored by the user, if canceling the subscription is needed.